docs/howto: combine multiple if clauses
This adds a Commented CUE that shows several examples of combining the
same set of if clauses.

Unification at the top-level of the evaluation is used in order to keep
the scope of the examples as general as possible, so that the reader can
more easily extract and use their lessons in both field and list
comprehensions - without having to first extract them from a
comprehension type that they're /not/ placed inside.

Closes cue-lang/docs-and-content#185.

Preview-Path: /docs/howto/combine-multiple-if-clauses/
Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>
Change-Id: I899019354b0b31c66dcff8a2458812011499f384
Dispatch-Trailer: {"type":"trybot","CL":1203075,"patchset":1,"ref":"refs/changes/75/1203075/1","targetBranch":"master"}