docs/howto: use the built-in function error
This adds a guide demonstrating how to use the built-in function
`error()` in a disjunction in order to emit custom error messages.

Discussion of the function's use in conditionals, such as

    if x == true {
    	error("x can't be true")
    }

... is left for later, when there will be more real-world, concrete
advice to offer around this potentially less-idiomatic CUE form.

Preview-Path: /docs/howto/use-the-built-in-function-error/
Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>
Change-Id: I93d6095f7127f59f3bd8cca98dbb20eba44b1e13
Dispatch-Trailer: {"type":"trybot","CL":1220330,"patchset":2,"ref":"refs/changes/30/1220330/2","targetBranch":"master"}